refactor(web): group all ConnectionForm fields validation in single module
Rollback the previous multi-file structure for connection form fields
and merge all related modules into a single `connectionFormFields.ts`
file.

This centralizes types, constants, default values, and validation in one
place, making it easy to see all form defaults and validation rules at a
glance. The organized sections and comments improve readability and
maintain existing behavior without needing to navigate multiple files.
