+++ web/src/components/network/connection-form/BridgeFields.test.tsx
@@ -24,8 +24,8 @@ import React from "react";
 import { screen } from "@testing-library/react";
 import { installerRender } from "~/test-utils";
 import { useAppForm } from "~/hooks/form";
-import { connectionFormOptions } from "./connectionValidation";
-import { BridgeStpMode } from "./bridgeFields";
+import { connectionFormOptions } from "./connectionFormFields";
+import { BridgeStpMode } from "./connectionFormFields";
 import { DeviceState } from "~/types/network";
 import { CONNECTION_TYPE } from "~/utils/network";
 import BridgeFields from "./BridgeFields";
